it's alive
Okay, this works. The allocation situation is still kind of a mess, but
I think I have enough escape hatches for usability's sake. Parsing
from a file is left as an exercise for the user (it's not hard). The
node manipulation routines will probably added on demand. Or maybe I
will get a wild hair and just do it at some point.